Move Generic to default type instance, and fill in rest of V1
